ai_service/requirements.txt
fastapi
uvicorn[standard]
numpy
scikit-learn
sentence-transformers
faiss-cpu
implicit 
scipy 
pandas
rapidfuzz
pymongo
certifi
geopy
xgboost 
scikit-learn 
pandas 
joblib
python-dotenv